# docs/pleya-protocol/v1/openapi.yaml
...
openapi: 3.1.0
info:
  title: Pleya Protocol
  version: "1.0.0"
  summary: Wire-contract tussen een Pleya-client en een Pleya Server.
  description: |
    Alles wat hier niet in staat, mag een client niet aannemen.

    Zes compatibiliteitsregels binnen v1:
      1. Een nieuw optioneel veld in een antwoord mag. Clients negeren
         onbekende velden.
      2. Een veld hernoemen of verwijderen mag niet.
      3. De betekenis van een bestaand veld wijzigen mag niet.
      4. Een nieuw verplicht veld in een aanvraag mag niet, in de query net zo
         min als in de body.
      5. Een aanvraagbody is gesloten (additionalProperties: false). Een client
         stuurt een nieuw optioneel veld pas wanneer capabilities of
         feature_level zegt dat de server het kent.
      6. Een nieuwe enum-waarde mag alleen waar x-unknown-safe true is. Elk
         enum-veld in dit document draagt die markering; check_protocol.sh
         weigert er een zonder.
  license:
    name: GPL-3.0-only
    identifier: GPL-3.0-only
servers:
  - url: /pleya/v1
    description: Het pad draagt de majorversie.
...
security:
  - bearerAuth: []
paths:
...
  /auth/api-tokens:
    post:
      tags: [auth]
      operationId: createApiToken
      summary: Maak een API-token.
      description: |
        Klasse `authenticated` op zichzelf, `admin` met `user_id` van een ander.
        Dezelfde regel als `GET /sessions`, en met opzet geen beheerroute: RB-20
        beschrijft een token dat een gebruiker voor zichzelf maakt, of een
        beheerder namens iemand.

        Een API-token is een sessie met `kind: api`. Het valt onder
        `DELETE /sessions/{id}`, onder dezelfde intrekking binnen twee seconden,
        en het staat in `GET /sessions` naast de toestellen.

        `secret` staat precies één keer in een antwoord: hier. De server bewaart
        alleen de hash ervan, dus een tweede kans is er niet.

        Het bereik komt nooit boven de rol van de eigenaar. Een aanvraag die dat
        wel doet krijgt `auth.scope_exceeds_role`, met het gevraagde bereik en de
        rol in `details`.

        Een aanvraag die zelf met een API-token is geauthenticeerd kan alleen
        tokens maken wanneer dat token bereik `admin` heeft; anders is het
        antwoord `404`, byte-gelijk aan dat van een beheerroute. Zonder die
        regel zou een `read`-token van een beheerder een `admin`-token kunnen
        maken, want het bereik wordt tegen de rol getoetst en de rol is dan
        `admin`.

        Beschikbaar wanneer `capabilities.api_tokens` waar is.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApiTokenRequest" }
      responses:
        "201":
          description: Het token, met het geheim erbij.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiTokenCreated" }
        "400":
          description: |
            auth.scope_exceeds_role wanneer het bereik boven de rol van de
            eigenaar uitkomt, session.invalid bij een naam, bereik of
            geldigheidsduur die niet klopt.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "401": { $ref: "#/components/responses/Error" }
        "404":
          description: |
            auth.user_not_found. user_id is niet van de aanvrager en de
            aanvrager is geen admin, of de aanvraag komt van een API-token
            zonder bereik `admin`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "500": { $ref: "#/components/responses/InternalError" }

    get:
      tags: [auth]
      operationId: listApiTokens
      summary: De levende API-tokens van een gebruiker.
      description: |
        Klasse `authenticated` op zichzelf, `admin` met `?user_id=`. Zonder
        geheimen: die bestaan na het aanmaken alleen nog bij de aanvrager.

        Ingetrokken en verlopen tokens staan er niet in, om dezelfde reden als
        bij `GET /sessions`: de lijst beantwoordt wat er nú met dit account kan
        praten.
      parameters:
        - name: user_id
          in: query
          required: false
          description: Alleen bruikbaar door `owner` en `admin`. Zonder deze parameter de eigen tokens.
          schema: { $ref: "#/components/schemas/Id" }
      responses:
        "200":
          description: De tokens.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiTokenList" }
        "401": { $ref: "#/components/responses/Error" }
        "404":
          description: auth.user_not_found. user_id is niet van de aanvrager en de aanvrager is geen admin.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Het accesstoken. Nooit in een querystring.
    streamSession:
      type: apiKey
      in: query
      name: ss
      description: |
        De niet-geheime helft van een browser-streamsessie. Het geheim reist in
        een cookie waarvan de naam deze id bevat, en de server valideert de twee
        samen. Alleen geldig op /stream, en alleen wanneer
        `capabilities.stream_sessions` waar is.
    streamToken:
      type: apiKey
      in: query
      name: stream_token
      description: |
        De enige uitzondering op de regel hierboven. Externe spelers kunnen geen
        header zetten. Geldig voor twee tot vijf minuten, gebonden aan een
        identiteit en een mediaresource, zonder rechten op de rest van de API.
...
  responses:
    Error:
      description: Fout met een stabiele machineleesbare code.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    InternalError:
      description: |
        `server.internal`. Elke operatie kan hem geven, want hij is het antwoord
        op een fout die de handler zelf niet had voorzien: de recovery-laag vangt
        een panic af en maakt er een envelop van in plaats van een verbroken
        verbinding of een leeg lichaam.

        `details.request_id` draagt dezelfde id als de logregel die de stack
        bevat. Dat is het enige dat een client ermee doet: tonen zodat een
        gebruiker hem kan doorgeven. De stack zelf verlaat de server nooit.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    Id:
      type: string
      minLength: 1
      description: |
        Ondoorzichtig voor de client. Er mag niets uit worden afgeleid, ook niet
        wanneer het op iets herkenbaars lijkt. Een pad is nooit een id.
    Timestamp:
      type: string
      format: date-time
      description: RFC 3339 in UTC, met Z-achtervoegsel.
...
    ErrorEnvelope:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code:
              type: string
              pattern: "^(auth|library|playback|session|settings|storage|server|job)\\.[a-z0-9_]+$"
              description: |
                Het contract. De client vertaalt codes en matcht nooit op message.

                Het domein is de helft voor de punt. De lijst is niet gesloten
                maar groeit per protocolvenster, en elke uitbreiding is een
                besluit: `server` en `settings` kwamen erbij met venster 1
                (DEC-135 en DEC-136). `job` kwam erbij met venster 2 (DEC-138),
                zodra S2.4 `job.not_cancellable` stuurde. Een client die een code
                niet kent valt terug op een generieke melding en takt niet op het domein.

                Codes die een client zelf verzint horen hier niet thuis. De
                webclient draagt `client.transport` en
                `client.malformed_response`; die komen nooit over de lijn en
                het contract keurt ze af.
            message:
              type: string
              description: Voor logs, niet voor de UI.
            retryable:
              type: boolean
              description: |
                Expliciet en geen afleiding uit de status, omdat een 503 soms wel
                en een 409 soms niet te herhalen is.
            details:
              type: object
              additionalProperties: true
...
    ApiTokenScope:
      type: string
      x-unknown-safe: true
      enum: [read, maintenance, admin]
      description: |
        Een ladder, laag naar hoog. `read` mag lezen wat de eigenaar mag lezen
        en niets beheren; `admin` haalt de adminklasse, mits de rol van de
        eigenaar dat ook doet. `maintenance` is gereserveerd voor de
        operationele handelingen die later komen (back-up, onderhoudsmodus,
        herstel) en onderscheidt zich vandaag in geen enkele operatie van
        `read`; dat staat hier expliciet in plaats van dat het gedrag het
        suggereert.
    ApiToken:
      type: object
      required: [id, user_id, name, scope, created_at, last_seen_at, expires_at]
      description: Een API-token zonder zijn geheim.
      properties:
        id: { $ref: "#/components/schemas/Id" }
        user_id: { $ref: "#/components/schemas/Id" }
        name:
          type: string
          maxLength: 64
          description: Het label in het overzicht. Geen document.
        scope: { $ref: "#/components/schemas/ApiTokenScope" }
        created_at: { $ref: "#/components/schemas/Timestamp" }
        last_seen_at: { $ref: "#/components/schemas/Timestamp" }
        expires_at: { $ref: "#/components/schemas/Timestamp" }
    ApiTokenList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ApiToken" }
    ApiTokenCreated:
      type: object
      required: [token, secret]
      description: |
        Een eigen schema en niet ApiToken met een optioneel geheimveld. Een
        optioneel `secret` op het lijsttype zou een lezer laten denken dat het
        er soms in kan staan, en dan is de vraag "kan dit antwoord een geheim
        dragen" niet meer aan het type te zien.
      properties:
        token: { $ref: "#/components/schemas/ApiToken" }
        secret:
          type: string
          description: |
            Het geheim, precies één keer. De server bewaart alleen de hash. Het
            begint met een herkenbaar voorvoegsel zodat het in een logbestand of
            een repository terug te vinden is.
    ApiTokenRequest:
      type: object
      additionalProperties: false
      required: [name, scope]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        scope: { $ref: "#/components/schemas/ApiTokenScope" }
        expires_in_days:
          type: integer
          minimum: 1
          maximum: 3650
          default: 90
          description: |
            Standaard negentig dagen. "Nooit" bestaat niet: een token zonder
            vervaldatum is de vorm die jaren later nog in een oude configuratie
            meeloopt en die niemand mist tot iemand hem vindt.
        user_id:
          allOf:
            - $ref: "#/components/schemas/Id"
          description: |
            Alleen bruikbaar door `owner` en `admin`. Zonder dit veld gaat het
            token over de aanvrager zelf.
...